fix: follow api routes bugs
